pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dictparser"
version = "0.1.6.dev"
description = "A Class serialization/deserialization library"
readme = "README.md"
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT No Attribution License (MIT-0)",
    "Operating System :: OS Independent",
]
dependencies = [
    "dataclasses; python_version<'3.7'",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/tgoncalves87/dictparser"
Issues = "https://github.com/tgoncalves87/dictparser/issues"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]

[tool.pylint.MAIN]
load-plugins = ["dictparser.pylint"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint."messages control"]
disable = [
    "no-else-return",
    "no-else-raise",
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring"
]

[tool.pyright]
useLibraryCodeForTypes=true
